Migrate line, chevron, arrow to Slang + vertex-buffer instancing
Completes the canvas-feature shader set on the new single-source path.
Line + chevron share lineInstance.slang; all four shaders pack color as a
single u32 (ABGR layout) end-to-end, eliminating the Uint8Array intermediate
and writeColorBytes/rgbaString helpers. Canvas2D rect/line/arrow fills use
abgrToCssRgba from the shared colorBits utility. Net -568 lines in the
canvas plugin.

+++ plugins/canvas/src/RenderFeatureDataRPC/rpcTypes.ts
@@ -55,14 +55,14 @@ export interface FeatureDataResult {
   // Connecting lines (introns) with strand info for dynamic chevron generation
   linePositions: Uint32Array
   lineYs: Float32Array
-  lineColors: Uint8Array
+  lineColors: Uint32Array
   lineDirections: Int8Array // strand direction: -1, 0, or 1
 
   // Strand arrows (at feature ends)
   arrowXs: Uint32Array
   arrowYs: Float32Array
   arrowDirections: Int8Array
-  arrowColors: Uint8Array
+  arrowColors: Uint32Array
 
   // Hit detection
   flatbushItems: FlatbushItem[]
